+++ src/core/object/watch/spec.js
@@ -718,7 +718,9 @@ describe('core/object/watch', () => {
 					spy(value, oldValue, info.path);
 				});
 
-				[...proxy][0].a++;
+				expect([...map]).toEqual([...new Map([[key, {a: 1}]])]);
+
+				[...proxy][0][1].a++;
 				expect(spy).toHaveBeenCalledWith(2, 1, [key, 'a']);
 				expect(map).toEqual(new Map([[key, {a: 2}]]));
 
@@ -903,6 +905,48 @@ describe('core/object/watch', () => {
 
 					proxy.a = 2;
 					expect(spy).toHaveBeenCalledWith(2, 1, ['a']);
+				}
+			});
+
+			it('deep setting of new properties', () => {
+				{
+					const
+						obj = {},
+						spy = jasmine.createSpy();
+
+					const {proxy, set} = watch(obj, {immediate: true, engine, deep: true}, (value, oldValue, info) => {
+						spy(value, oldValue, info.path);
+					});
+
+					set('a.b', 1);
+					expect(spy).toHaveBeenCalledWith(1, undefined, ['a', 'b']);
+					expect(proxy.a?.b).toBe(1);
+					expect(obj.a?.b).toBe(1);
+
+					proxy.a.b = 2;
+					expect(spy).toHaveBeenCalledWith(2, 1, ['a', 'b']);
+					expect(proxy.a?.b).toBe(2);
+					expect(obj.a?.b).toBe(2);
+				}
+
+				{
+					const
+						obj = {},
+						spy = jasmine.createSpy();
+
+					const {proxy} = watch(obj, {immediate: true, engine, deep: true}, (value, oldValue, info) => {
+						spy(value, oldValue, info.path);
+					});
+
+					set(proxy, 'a.b', 1, engine);
+					expect(spy).toHaveBeenCalledWith(1, undefined, ['a', 'b']);
+					expect(proxy.a?.b).toBe(1);
+					expect(obj.a?.b).toBe(1);
+
+					proxy.a.b = 2;
+					expect(spy).toHaveBeenCalledWith(2, 1, ['a', 'b']);
+					expect(proxy.a?.b).toBe(2);
+					expect(obj.a?.b).toBe(2);
 				}
 			});