+++ CHANGELOG.md
@@ -15,7 +15,19 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :boom: Breaking Change
 
-* By default, the `ES` flag is set to `ES2021` for development builds
+* By default, the `ES` flag is set to `ES2021` for development builds `config`
+
+#### :bug: Bug Fix
+
+* `core/object/watch`:
+  * Fixed setting new properties of a watchable object by a deep path using the accessor-based engine
+  * Now watchable Map uses the `entries` iterator by default
+
+## v3.87.0 (2022-08-08)
+
+#### :rocket: New Feature
+
+* Added a new overload for the `intoIter` method to create an iterator from an array-like object
 
 ## v3.86.4 (2022-07-05)